feat : 답변 완료된 문의 삭제 차단

+++ Mavis-Api/src/main/java/com/mavis/api/inquiry/service/InquiryService.java
@@ -10,6 +10,8 @@ import com.mavis.api.inquiry.implement.InquiryReader;
 import com.mavis.common.util.DateFormatters;
 import com.mavis.domain.domains.inquiry.domain.Inquiry;
 import com.mavis.domain.domains.inquiry.domain.InquiryAnswer;
+import com.mavis.domain.domains.inquiry.exception.InquiryAlreadyAnsweredCannotDeleteException;
+import com.mavis.domain.domains.inquiry.exception.InquiryAlreadyAnsweredCannotDeleteException;
 import com.mavis.domain.domains.inquiry.exception.UnauthorizedInquiryException;
 import com.mavis.domain.domains.inquiry.repository.InquiryRepository;
 import com.mavis.domain.domains.product.domain.Product;
@@ -46,6 +48,9 @@ public class InquiryService {
         Inquiry inquiry = inquiryReader.findById(inquiryId);
         if (!inquiry.getUser().getId().equals(user.getId())) {
             throw UnauthorizedInquiryException.EXCEPTION;
+        }
+        if (inquiry.getInquiryAnswer() != null && !inquiry.getInquiryAnswer().isDeleted()) {
+            throw InquiryAlreadyAnsweredCannotDeleteException.EXCEPTION;
         }
         inquiry.delete();
     }
